added View.prototype.render in specific file view.js
